micropython/requirements.txt
adafruit-ampy==1.1.0
altgraph==0.17.4
beautifulsoup4==4.12.3
bitarray==2.9.2
bitstring==4.2.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
ecdsa==0.19.0
esptool==4.7.0
idna==3.10
intelhex==2.3.0
mpfshell==0.9.3
packaging==24.1
pefile==2024.8.26
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyserial==3.5
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
reedsolo==1.7.0
requests==2.32.3
setuptools==75.0.0
six==1.16.0
soupsieve==2.6
urllib3==2.2.3
websocket-client==1.8.0
